// frontend/src/app/shared/models/establecimiento.model.ts
export interface Establecimiento {
  idEstab?: string;
  idCat: number;
  idSubcat?: number;
  nombre: string;
  direccion?: string;
  telFijo?: string;
  telMovil?: string;
  whatsapp?: string;
  correo?: string;
  web?: string;
  instagram?: string;
  facebook?: string;
  anioInauguracion?: number;
  estado: string;
  personalPermanente?: number;
  personalEventual?: number;
  redesSociales?: string;
  observaciones?: string;
}